requirements.txt
# Core RAG framework
langchain>=0.2.0
langchain-community>=0.2.0

# Vector databases
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Embeddings & LLM utilities
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.1.0

# Tokenization & text utils
tiktoken>=0.6.0

# Document loaders
pypdf>=4.0.0

# Environment variables
python-dotenv>=1.0.1

# Optional (but recommended for eval stability)
scikit-learn>=1.4.0
numpy>=1.26.0
